Add files via upload
When a bank considers whether to approve a loan applicant, one of the most important
factors for the bank to consider is whether that applicant will be a “bad” credit risk and default
on the loan. I will use previous applicant data titled, “Equity.JMP”, in order to create a model to
better predict the credit risk of a loan applicant.